hotplug: add a up_frequency tunable
One of the main design goals of this driver was to provide
a generic approach of hotplugging. It turned out that powerful
cpus are unlikely to go all the way up to their max frequency
when under load.
This might be dua a wrongly configured cpu governor or just
because the device is too fast.
Adding a up_frequency tunable provides a simple a way to define
the frequency when the hotplug driver should care to online cpus.

By default, this is the max frequency reported by the cpu policy,
but can be changed to anything else.
One nice side-effect is that we only get the actual policy ones
instead of each sampling rate.

Signed-off-by: Blechd0se <alex.christ@hotmail.de>